+++ src/graphql/node/crud.rs
@@ -4,8 +4,7 @@ use crate::graphql::{customer::broadcast_customer_networks, get_customer_network
 
 use super::{
     super::{Role, RoleGuard},
-    Node, NodeInput, NodeMutation, NodeQuery, NodeTotalCount, PortNumber, ServerAddress,
-    ServerPort, Setting,
+    Node, NodeInput, NodeMutation, NodeQuery, NodeSetting, NodeTotalCount, PortNumber,
 };
 use async_graphql::{
     connection::{query, Connection, EmptyFields},
@@ -15,10 +14,7 @@ use async_graphql::{
 use bincode::Options;
 use chrono::Utc;
 use review_database::{Indexed, IterableMap, Store};
-use std::{
-    collections::HashMap,
-    net::{IpAddr, Ipv4Addr, SocketAddr},
-};
+use std::{collections::HashMap, net::IpAddr};
 use tracing::error;
 
 #[Object]
@@ -58,6 +54,19 @@ impl NodeQuery {
         Ok(bincode::DefaultOptions::new()
             .deserialize(value.as_ref())
             .map_err(|_| "invalid value in database")?)
+    }
+}
+
+fn parse_str_to_ip<'em>(
+    ip_str: Option<&str>,
+    error_message: &'em str,
+) -> Result<Option<IpAddr>, &'em str> {
+    match ip_str {
+        Some(ip_str) => ip_str
+            .parse::<IpAddr>()
+            .map(Some)
+            .map_err(|_| error_message),
+        None => Ok(None),
     }
 }
 
@@ -127,143 +136,88 @@ impl NodeMutation {
                 .as_str()
                 .parse::<u32>()
                 .map_err(|_| "invalid customer ID")?;
-            let piglet_giganto_ip = if let Some(ip) = piglet_giganto_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: storage")?,
-                )
-            } else {
-                None
-            };
-            let piglet_review_ip = if let Some(ip) = piglet_review_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: administration")?,
-                )
-            } else {
-                None
-            };
-            let giganto_ingestion_ip = if let Some(ip) = giganto_ingestion_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: receiving")?,
-                )
-            } else {
-                None
-            };
-            let giganto_publish_ip = if let Some(ip) = giganto_publish_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: sending")?,
-                )
-            } else {
-                None
-            };
-            let giganto_graphql_ip = if let Some(ip) = giganto_graphql_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: web")?,
-                )
-            } else {
-                None
-            };
-            let reconverge_review_ip = if let Some(ip) = reconverge_review_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: administration")?,
-                )
-            } else {
-                None
-            };
-            let reconverge_giganto_ip = if let Some(ip) = reconverge_giganto_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: storage")?,
-                )
-            } else {
-                None
-            };
-            let hog_review_ip = if let Some(ip) = hog_review_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: administration")?,
-                )
-            } else {
-                None
-            };
-            let hog_giganto_ip = if let Some(ip) = hog_giganto_ip {
-                Some(
-                    ip.as_str()
-                        .parse::<IpAddr>()
-                        .map_err(|_| "invalid IP address: storage")?,
-                )
-            } else {
-                None
-            };
 
             let value = Node {
                 id: u32::MAX,
-                customer_id,
-                name,
-                description,
-                hostname,
-
-                review,
-                review_port,
-                review_web_port,
-
-                piglet,
-                piglet_giganto_ip,
-                piglet_giganto_port,
-                piglet_review_ip,
-                piglet_review_port,
-                save_packets,
-                http,
-                office,
-                exe,
-                pdf,
-                html,
-                txt,
-                smtp_eml,
-                ftp,
-
-                giganto,
-                giganto_ingestion_ip,
-                giganto_ingestion_port,
-                giganto_publish_ip,
-                giganto_publish_port,
-                giganto_graphql_ip,
-                giganto_graphql_port,
-                retention_period,
-
-                reconverge,
-                reconverge_review_ip,
-                reconverge_review_port,
-                reconverge_giganto_ip,
-                reconverge_giganto_port,
-
-                hog,
-                hog_review_ip,
-                hog_review_port,
-                hog_giganto_ip,
-                hog_giganto_port,
-                protocols,
-                protocol_list,
-                sensors,
-                sensor_list,
-
                 creation_time: Utc::now(),
+                as_is: None,
+                to_be: Some(NodeSetting {
+                    name,
+                    customer_id,
+                    description,
+                    hostname,
 
-                apply_target_id: None,
-                apply_in_progress: false,
+                    review,
+                    review_port,
+                    review_web_port,
+
+                    piglet,
+                    piglet_giganto_ip: parse_str_to_ip(
+                        piglet_giganto_ip.as_deref(),
+                        "invalid IP address: storage",
+                    )?,
+                    piglet_giganto_port,
+                    piglet_review_ip: parse_str_to_ip(
+                        piglet_review_ip.as_deref(),
+                        "invalid IP address: administration",
+                    )?,
+                    piglet_review_port,
+                    save_packets,
+                    http,
+                    office,
+                    exe,
+                    pdf,
+                    html,
+                    txt,
+                    smtp_eml,
+                    ftp,
+
+                    giganto,
+                    giganto_ingestion_ip: parse_str_to_ip(
+                        giganto_ingestion_ip.as_deref(),
+                        "invalid IP address: receiving",
+                    )?,
+                    giganto_ingestion_port,
+                    giganto_publish_ip: parse_str_to_ip(
+                        giganto_publish_ip.as_deref(),
+                        "invalid IP address: sending",
+                    )?,
+                    giganto_publish_port,
+                    giganto_graphql_ip: parse_str_to_ip(
+                        giganto_graphql_ip.as_deref(),
+                        "invalid IP address: web",
+                    )?,
+                    giganto_graphql_port,
+                    retention_period,
+
+                    reconverge,
+                    reconverge_review_ip: parse_str_to_ip(
+                        reconverge_review_ip.as_deref(),
+                        "invalid IP address: administration",
+                    )?,
+                    reconverge_review_port,
+                    reconverge_giganto_ip: parse_str_to_ip(
+                        reconverge_giganto_ip.as_deref(),
+                        "invalid IP address: storage",
+                    )?,
+                    reconverge_giganto_port,
+
+                    hog,
+                    hog_review_ip: parse_str_to_ip(
+                        hog_review_ip.as_deref(),
+                        "invalid IP address: administration",
+                    )?,
+                    hog_review_port,
+                    hog_giganto_ip: parse_str_to_ip(
+                        hog_giganto_ip.as_deref(),
+                        "invalid IP address: storage",
+                    )?,
+                    hog_giganto_port,
+                    protocols,
+                    protocol_list,
+                    sensors,
+                    sensor_list,
+                }),
             };
             let id = map.insert(value)?;
             (id, customer_id)
@@ -317,34 +271,30 @@ impl NodeMutation {
         old: NodeInput,
         new: NodeInput,
     ) -> Result<ID> {
-        let (review, customer_is_changed, customer_id) = {
-            let i = id.as_str().parse::<u32>().map_err(|_| "invalid ID")?;
-
-            let store = crate::graphql::get_store(ctx).await?;
-            let map = store.node_map();
-            map.update(i, &old, &new)?;
-
-            (
-                new.review,
-                new.customer_id != old.customer_id,
-                new.customer_id,
-            )
-        };
-
-        if review && customer_is_changed {
-            if let Ok(customer_id) = customer_id.as_str().parse::<u32>() {
-                let store = crate::graphql::get_store(ctx).await?;
-
-                if let Ok(networks) = get_customer_networks(&store, customer_id) {
-                    if let Err(e) = broadcast_customer_networks(ctx, &networks).await {
-                        error!("failed to broadcast internal networks. {e:?}");
-                    }
-                }
-            }
+        if !validate_update_input(&old, &new) {
+            return Err("Invalid combination of old and new values".into());
         }
 
+        let i = id.as_str().parse::<u32>().map_err(|_| "invalid ID")?;
+        let store = crate::graphql::get_store(ctx).await?;
+        let map = store.node_map();
+        map.update(i, &old, &new)?;
         Ok(id)
     }
+}
+
+fn validate_update_input(old: &NodeInput, new: &NodeInput) -> bool {
+    matches!(
+        (
+            old.as_is.as_ref(),
+            old.to_be.as_ref(),
+            new.as_is.as_ref(),
+            new.to_be.as_ref(),
+        ),
+        (None, Some(_), None, Some(_))
+            | (Some(_), None, Some(_), Some(_))
+            | (Some(_), Some(_), Some(_), _)
+    )
 }
 
 async fn load(
@@ -359,118 +309,6 @@ async fn load(
     super::super::load(&map, after, before, first, last, NodeTotalCount)
 }
 
-/// Returns the node settings.
-///
-/// # Errors
-///
-/// Returns an error if the node settings could not be retrieved.
-pub fn get_node_settings(db: &Store) -> Result<Vec<Setting>> {
-    let map = db.node_map();
-    let mut output = Vec::new();
-    for (_key, value) in map.iter_forward()? {
-        let node = bincode::DefaultOptions::new()
-            .deserialize::<Node>(value.as_ref())
-            .map_err(|_| "invalid value in database")?;
-
-        let piglet: Option<ServerAddress> = if node.piglet {
-            Some(ServerAddress {
-                web: None,
-                rpc: Some(SocketAddr::new(
-                    node.piglet_review_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.piglet_review_port.unwrap_or_default(),
-                )),
-                public: Some(SocketAddr::new(
-                    node.piglet_giganto_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.piglet_giganto_port.unwrap_or_default(),
-                )),
-                ing: Some(SocketAddr::new(IpAddr::V4(Ipv4Addr::UNSPECIFIED), 0)),
-            })
-        } else {
-            None
-        };
-        let giganto = if node.giganto {
-            Some(ServerAddress {
-                web: Some(SocketAddr::new(
-                    node.giganto_graphql_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.giganto_graphql_port.unwrap_or_default(),
-                )),
-                rpc: None,
-                public: Some(SocketAddr::new(
-                    node.giganto_publish_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.giganto_publish_port.unwrap_or_default(),
-                )),
-                ing: Some(SocketAddr::new(
-                    node.giganto_ingestion_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.giganto_ingestion_port.unwrap_or_default(),
-                )),
-            })
-        } else {
-            None
-        };
-
-        let review = if node.review {
-            Some(ServerPort {
-                rpc_port: node.review_port.unwrap_or_default(),
-                web_port: node.review_web_port.unwrap_or_default(),
-            })
-        } else {
-            None
-        };
-        let reconverge = if node.reconverge {
-            Some(ServerAddress {
-                web: None,
-                rpc: Some(SocketAddr::new(
-                    node.reconverge_review_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.reconverge_review_port.unwrap_or_default(),
-                )),
-                public: Some(SocketAddr::new(
-                    node.reconverge_giganto_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.reconverge_giganto_port.unwrap_or_default(),
-                )),
-                ing: None,
-            })
-        } else {
-            None
-        };
-        let hog = if node.hog {
-            Some(ServerAddress {
-                web: None,
-                rpc: Some(SocketAddr::new(
-                    node.hog_review_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.hog_review_port.unwrap_or_default(),
-                )),
-                public: Some(SocketAddr::new(
-                    node.hog_giganto_ip
-                        .unwrap_or_else(|| IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-                    node.hog_giganto_port.unwrap_or_default(),
-                )),
-                ing: None,
-            })
-        } else {
-            None
-        };
-
-        output.push(Setting {
-            name: node.hostname,
-            piglet,
-            giganto,
-            hog,
-            reconverge,
-            review,
-        });
-    }
-
-    Ok(output)
-}
-
 /// Returns the customer id of review node.
 ///
 /// # Errors
@@ -483,8 +321,11 @@ pub fn get_customer_id_of_review_host(db: &Store) -> Result<Option<u32>> {
         let node = bincode::DefaultOptions::new()
             .deserialize::<Node>(value.as_ref())
             .map_err(|_| "invalid value in database")?;
-        if node.review {
-            return Ok(Some(node.customer_id));
+
+        if let Some(as_is) = &node.as_is {
+            if as_is.review {
+                return Ok(Some(as_is.customer_id));
+            }
         }
     }
     Ok(None)